+++ frontend/src/pages/organizations/index.tsx
@@ -1,6 +1,6 @@
 import { createGetServerSideProps } from "core/helpers/page";
 import { NextPageWithLayout } from "core/helpers/types";
-import OrganizationsLayout from "./OrganizationsLayout";
+import OrganizationsLayout from "../../organizations/layouts/OrganizationsLayout";
 import { useTranslation } from "next-i18next";
 import Page from "core/components/Page";
 import {
